# swagger/RajBhuva.postman_collection.json-OpenApi3Json.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "RajBhuva",
    "contact": {},
    "version": "1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "variables": {}
    }
  ],
  "paths": {
...
    "/product/edit/660b03134fbf9df2b1f2abb3": {
      "put": {
        "tags": [
          "product"
        ],
        "summary": "edit data",
        "operationId": "editdata",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/editdatarequest"
                  },
                  {
                    "example": {
                      "product_name": "apple",
                      "price": 705,
                      "descrption": "apple labore dol"
                    }
                  }
                ]
              },
              "example": {
                "product_name": "apple",
                "price": 705,
                "descrption": "apple labore dol"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/student/delete/660c2e437248711c75c1f0e8": {
      "delete": {
        "tags": [
          "student"
        ],
        "summary": "delete data",
        "operationId": "deletedata",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false,
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/emailverfication": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "emailVerfication",
        "operationId": "emailVerfication",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/emailVerficationRequest"
                  },
                  {
                    "example": {
                      "email": "rkbhuva569@gmail.com"
                    }
                  }
                ]
              },
              "example": {
                "email": "rkbhuva569@gmail.com"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    },
    "/api/resetpassword": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "resetpassword",
        "operationId": "resetpassword",
        "parameters": [],
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/resetpasswordrequest"
                  },
                  {
                    "example": {
                      "email": "rkbhuva569@gmail.com",
                      "password": "456",
                      "confirmpassword": "456"
                    }
                  }
                ]
              },
              "example": {
                "email": "rkbhuva569@gmail.com",
                "password": "456",
                "confirmpassword": "456"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "headers": {}
          }
        },
        "deprecated": false
      }
    }
  },
  "components": {
    "schemas": {
...
      "editdatarequest": {
        "title": "editdatarequest",
        "required": [
          "product_name",
          "price",
          "descrption"
        ],
        "type": "object",
        "properties": {
          "product_name": {
            "type": "string"
          },
          "price": {
            "type": "integer",
            "format": "int32"
          },
          "descrption": {
            "type": "string"
          }
        },
        "example": {
          "product_name": "apple",
          "price": 705,
          "descrption": "apple labore dol"
        }
      },
...
      "emailVerficationRequest": {
        "title": "emailVerficationRequest",
        "required": [
          "email"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          }
        },
        "example": {
          "email": "rkbhuva569@gmail.com"
        }
      },
      "resetpasswordrequest": {
        "title": "resetpasswordrequest",
        "required": [
          "email",
          "password",
          "confirmpassword"
        ],
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "confirmpassword": {
            "type": "string"
          }
        },
        "example": {
          "email": "rkbhuva569@gmail.com",
          "password": "456",
          "confirmpassword": "456"
        }
      }
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [],
...
}